LayoutTests/fast/text/thai-joiner.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<style>
@font-face {
    font-family: "WebFont";
    src: url("resources/Thai_test.svg") format("svg");
}
</style>
</head>
<body>
<div style="font: 48px 'WebFont';"><span id="target">&#x0e2a;&#x200d;&#x0e42;</span></div>
<script>
window.jsTestIsAsync = true;
[...document.fonts.entries()][0].load().then(function() {
    shouldBeGreaterThanOrEqual("50", "document.getElementById('target').offsetWidth");
    finishJSTest();
});
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
