LayoutTests/fast/text/thai-joiner-expected.txt

PASS 50 is >= document.getElementById('target').offsetWidth
PASS successfullyParsed is true

TEST COMPLETE
ส‍โ
